+++ src/validator.js
@@ -47,13 +47,10 @@ let validate = ( xformStr, options = {} ) => {
     if ( xform ) {
         xform.checkStructure( warnings, errors );
         xform.checkBinds( warnings, errors );
+        xform.checkAppearances( warnings, errors );
 
         if ( options.openclinica ) {
             xform.checkOpenClinicaRules( warnings, errors );
-            // OpenClinica would like all appearance warnings to be output as errors, for now
-            xform.checkAppearances( errors, errors );
-        } else {
-            xform.checkAppearances( warnings, errors );
         }
     }